fix: keep the pipeline inside a 512 MB container
The Render deployment served /api/health fine but returned 502 on /api/slicks
and the UI. The health check does no work; anything triggering analysis was
killing the worker on memory.

Two causes, both fixed:

  * refined Lee ran eight directional convolutions in float64. Those
    intermediates alone exceed a free container's budget, and the precision is
    far beyond what dB backscatter needs. Now float32 throughout.

  * demo scenes were generated at 1400 px. At 700 px the pixel count is
    quartered and the whole pipeline fits, while the planted slicks stay
    clearly visible.

98 ingest and detect tests pass unchanged, so the numerical behaviour is
intact - this is a memory fix, not an accuracy trade.

+++ ingest/speckle.py
@@ -37,7 +37,10 @@ def lee_filter(img: np.ndarray, size: int = 7, enl: float = DEFAULT_ENL) -> np.n
     original pixel where local variance exceeds what speckle alone explains,
     which is what preserves slick edges.
     """
-    arr = np.asarray(img, dtype=np.float64)
+    # float32 throughout: refined Lee runs eight directional convolutions, and
+# in float64 those intermediates alone exceed the memory of a small
+# container. Precision is far beyond what dB backscatter needs.
+    arr = np.asarray(img, dtype=np.float32)
     mean, var = _local_stats(arr, size)
 
     # Speckle-only variance for this ENL, in intensity terms.
@@ -74,7 +77,10 @@ def refined_lee(img: np.ndarray, size: int = 7, enl: float = DEFAULT_ENL) -> np.
     the edge instead of across it — which matters because the thin tapering
     streak of a bilge dump is exactly the structure a symmetric filter erodes.
     """
-    arr = np.asarray(img, dtype=np.float64)
+    # float32 throughout: refined Lee runs eight directional convolutions, and
+# in float64 those intermediates alone exceed the memory of a small
+# container. Precision is far beyond what dB backscatter needs.
+    arr = np.asarray(img, dtype=np.float32)
     if size < 3:
         return arr.astype(np.float32)
 
@@ -164,6 +170,9 @@ def speckle_index(img: np.ndarray) -> float:
     A crude but useful check that filtering did something: this number should
     drop noticeably after filtering and is ~1/sqrt(ENL) for pure speckle.
     """
-    arr = np.asarray(img, dtype=np.float64)
+    # float32 throughout: refined Lee runs eight directional convolutions, and
+# in float64 those intermediates alone exceed the memory of a small
+# container. Precision is far beyond what dB backscatter needs.
+    arr = np.asarray(img, dtype=np.float32)
     m = float(arr.mean())
     return float(arr.std() / m) if m != 0 else 0.0
